Unify aggregated and non-aggregated range proofs.
This removes the `AggregatedProof` struct entirely and moves its validation
logic into the `RangeProof` struct, which now handles both aggregated and
unaggregated proofs.  The former `RangeProof::verify` is now
`RangeProof::verify_single`, a convenience wrapper.
